# docs/SERVER_MODE/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MRSS is a modern, cross-platform desktop RSS reader with auto-translation, smart feed discovery, and AI-powered summarization.",
        "title": "MRSS API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/marcomarcogd/MRSS/issues"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "1.4.2"
    },
    "host": "localhost:1234",
    "basePath": "/api",
    "paths": {
...
        "/freshrss/sync": {
            "post": {
                "description": "Perform bidirectional synchronization with FreshRSS server (pull and push changes)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "freshrss"
                ],
                "summary": "Sync with FreshRSS",
                "responses": {
                    "200": {
                        "description": "Sync started status (status, message)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request (FreshRSS disabled or incomplete settings)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/freshrss/sync-feed": {
            "post": {
                "description": "Synchronize articles for a specific FreshRSS feed/stream",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "freshrss"
                ],
                "summary": "Sync single FreshRSS feed",
                "parameters": [
                    {
                        "type": "string",
                        "description": "FreshRSS stream ID",
                        "name": "stream_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sync started status (status, message)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request (FreshRSS disabled or stream_id missing)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/network/detect": {
            "post": {
                "description": "Detect network speed and latency, then optimize max concurrent refreshes setting",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "network"
                ],
                "summary": "Detect network speed",
                "responses": {
                    "200": {
                        "description": "Network detection result (speed_level, bandwidth_mbps, latency_ms, max_concurrency)",
                        "schema": {
                            "$ref": "#/definitions/network.DetectionResult"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/opml/export": {
            "get": {
                "description": "Export all feeds as an OPML file (server mode - direct download)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/xml"
                ],
                "tags": [
                    "opml"
                ],
                "summary": "Export OPML file",
                "responses": {
                    "200": {
                        "description": "OPML XML file",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/opml/import": {
            "post": {
                "description": "Import feeds from an OPML file (server mode - requires file upload)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "opml"
                ],
                "summary": "Import OPML file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "OPML file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Import result (success, imported, total)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request (invalid file or format)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "network.DetectionResult": {
            "type": "object",
            "properties": {
                "bandwidth_mbps": {
                    "type": "number"
                },
                "detection_success": {
                    "type": "boolean"
                },
                "detection_time": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "latency_ms": {
                    "type": "integer"
                },
                "max_concurrency": {
                    "type": "integer"
                },
                "speed_level": {
                    "$ref": "#/definitions/network.SpeedLevel"
                }
            }
        },
        "network.SpeedLevel": {
            "type": "string",
            "enum": [
                "slow",
                "medium",
                "fast"
            ],
            "x-enum-varnames": [
                "SpeedSlow",
                "SpeedMedium",
                "SpeedFast"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}